Extend struct cmode_'s value from int to long to make additional space for channel modes for some ircds.

+++ include/channels.h
@@ -73,7 +73,7 @@ struct chanban_
 struct cmode_
 {
         char mode;
-        unsigned int value;
+        unsigned long value;
 };
 
 struct extmode
